+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 ## Unreleased
 * Fix serialization of `ThreeDSecureLookupAdditionalInformation`
 * Return `AuthenticationInsight` when requested on `PaymentMethodNonce.create`
+* Add `roomTax` field to `TransactionIndustryDataRequest`
+* Add `noShow` field to `TransactionIndustryDataRequest`
+* Add `advancedDeposit` field to `TransactionIndustryDataRequest`
+* Add `fireSafe` field to `TransactionIndustryDataRequest`
+* Add `propertyPhone` field to `TransactionIndustryDataRequest`
+* Add `additionalChargeRequests` field to `TransactionIndustryDataRequest`
 
 ## 2.96.0
 * Add `ThreeDSecureLookup`